feat(i18n): updated the JavaFX .fxml configuration files for internationalization
What:
updated the JavaFX .fxml configuration files to use the added translation keys instead of labels in only English.

Why:
This is another (albeit temporary)step towards internationalizing the application, namely, having all application labels and buttons (be able to be) translated to a certain language.

Note:
The .fxml configurations will be later updated to use StringProperties of controllers, as opposed to hard-coded strings, so as to enable dynamically changing the labels.